requirements.txt
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
cssselect==1.2.0
deepl==1.11.0
et-xmlfile==1.1.0
exceptiongroup==1.0.4
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.8.2
google-auth==2.15.0
google-auth-oauthlib==0.8.0
gspread==5.7.2
gspread-formatting==1.1.2
h11==0.14.0
html2text==2020.1.16
httplib2==0.21.0
idna==3.4
jieba3k==0.35.1
Jinja2==3.1.2
joblib==1.2.0
loguru==0.6.0
lxml==4.9.1
markdownify==0.11.6
MarkupSafe==2.1.1
newspaper3k==0.2.8
nltk==3.8
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.0.10
outcome==1.2.0
Pillow==9.3.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.2
Pygments==2.13.0
pyparsing==3.0.9
PySimpleGUI==4.60.5
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
PyYAML==6.0
readability-lxml==0.8.1
regex==2022.10.31
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
rich==12.6.0
rsa==4.9
selenium==4.7.2
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
stackprinter==0.2.10
tinysegmenter==0.3
tldextract==3.4.0
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
urllib3==1.26.13
wget==3.2
wsproto==1.2.0
